# order-service/docs/swagger.json
{
  "swagger": "2.0",
  "basePath": "/api/v1",
  "host": "localhost:3000",
  "info": {
    "title": "Order Service API",
    "description": "API documentation for Order Service",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    }
  },
  "paths": {
    "/inventory/{product_id}/{warehouse_id}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Get inventory for a product",
        "description": "Returns current inventory level for a specific product at a warehouse",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "warehouse_id",
            "in": "path",
            "description": "Warehouse ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.InventoryResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/inventory/batch": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Get inventory for multiple products",
        "description": "Returns current inventory levels for multiple products across warehouses",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Inventory query request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.InventoryBatchRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.InventoryBatchResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/inventory/confirm": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Confirm stock deduction for an order",
        "description": "Confirms the deduction of reserved stock for an order (typically after payment). Calls the warehouse service endpoint at /api/v1/inventory/commit.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Stock operation request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.StockOperationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.StockOperationResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/inventory/release": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Release a stock reservation",
        "description": "Releases a previously created stock reservation, making the stock available again. Calls the warehouse service endpoint at /api/v1/inventory/cancel.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Stock operation request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.StockOperationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.StockOperationResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/inventory/reserve": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Reserve stock for an order",
        "description": "Creates stock reservations for items in an order. Uses either synchronous HTTP calls or asynchronous messaging based on configuration.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Stock reservation request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.StockReservationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.StockReservationResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "model.InventoryBatchRequest": {
      "type": "object",
      "required": [
        "items"
      ],
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.InventoryQuery"
          }
        }
      }
    },
    "model.InventoryBatchResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/model.InventoryResponse"
          }
        }
      }
    },
    "model.InventoryQuery": {
      "type": "object",
      "required": [
        "product_id",
        "warehouse_id"
      ],
      "properties": {
        "product_id": {
          "type": "integer"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
    "model.InventoryResponse": {
      "type": "object",
      "properties": {
        "available_quantity": {
          "type": "integer"
        },
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        },
        "reserved_quantity": {
          "type": "integer"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
    "model.OrderItemRequest": {
      "type": "object",
      "required": [
        "product_id",
        "quantity",
        "unit_price",
        "warehouse_id"
      ],
      "properties": {
        "order_id": {
          "type": "integer",
          "description": "Added for compatibility with warehouse service"
        },
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer",
          "minimum": 1
        },
        "unit_price": {
          "type": "number",
          "minimum": 0
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
...
    "model.StockOperationRequest": {
      "type": "object",
      "required": [
        "order_id"
      ],
      "properties": {
        "order_id": {
          "type": "integer"
        },
        "reservation_id": {
          "type": "string"
        }
      }
    },
    "model.StockOperationResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
    "model.StockReservationItem": {
      "type": "object",
      "properties": {
        "available": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "product_id": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
    "model.StockReservationRequest": {
      "type": "object",
      "required": [
        "items",
        "order_id"
      ],
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.OrderItemRequest"
          }
        },
        "order_id": {
          "type": "integer"
        },
        "reserve_until": {
          "type": "string"
        }
      }
    },
    "model.StockReservationResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.StockReservationItem"
          }
        },
        "message": {
          "type": "string"
        },
        "order_id": {
          "type": "integer"
        },
        "reservation_id": {
          "type": "string"
        },
        "reserved_until": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
...
    "response.ErrorInfo": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "response.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/response.ErrorInfo"
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
...
  }
}